Reference

Generate a world and run actions

Every program must start and end with calls to these functions.

generate_world()

Create Karel's world

run_actions()

Run actions

Available actions for Karel

These functions implement the four basic actions that Karel can perform (move, turn left, put and pick beepers). If Karel’s superpowers are turned on, she can also turn right and turn around.

move() turn_left() put_beeper() pick_beeper() turn_right() turn_around()

Available actions for Karel

load_super_karel()

Turn on Karel's superpowers

Conditions that Karel can test

These group of functions return a logical value TRUE or FALSE according to Karel’s evaluation of her world.

front_is_clear() front_is_blocked() left_is_clear() left_is_blocked() right_is_clear() right_is_blocked() beepers_present() no_beepers_present() karel_has_beepers() karel_has_no_beepers() facing_east() facing_west() facing_north() facing_south()

Conditions that Karel can test

Internal functions

Selected internal functions that could be useful to use the package in a more advanced way.

get_pkg_env()

Get Karel's environment

plot_static_world()

Plot the world at a given time