Referencias

Generar un mundo y ejecutar acciones

Todo programa debe comenzar y terminar con estas funciones.

generar_mundo()

Generar el mundo de Karel

ejecutar_acciones()

Ejecutar acciones

Acciones que Karel puede realizar

Funciones que implementan las cuatro actividades básicas que Karel sabe realizar (avanzar, girar a la izquierda, poner y juntar “cosos”). Si se habilitan los superpoderes de Karel, también puede girar a la derecha y darse vuelta.

avanzar() girar_izquierda() poner_coso() juntar_coso() girar_derecha() darse_vuelta()

Acciones que Karel puede realizar

cargar_super_karel()

Habilitar los superpoderes de Karel

Condiciones que Karel puede verificar

Este conjunto de funciones devuelven un valor lógico TRUE o FALSE según la evaluación que Karel puede hacer de su mundo.

frente_abierto() frente_cerrado() izquierda_abierto() izquierda_cerrado() derecha_abierto() derecha_cerrado() hay_cosos() no_hay_cosos() karel_tiene_cosos() karel_no_tiene_cosos() mira_al_este() mira_al_norte() mira_al_oeste() mira_al_sur()

Condiciones que Karel puede verificar

Funciones internas

Selección de funciones internas que podrían ser útiles para un uso algo más avanzado del paquete.

get_pkg_env()

Get Karel's environment

plot_static_world()

Plot the world at a given time