Plot base world

This functions produces the initial plot of the world, with its size and all the walls if there are any. It doesn't plot Karel or the beepers, since those things can change with time.

Uso

plot_base_world()

Valor

A ggplot with the world